Split the backend tests into their own separate file
- The intention is to have one test file for each endpoint
